system/pdf_recognizer/requirements.txt
marker-pdf
torch
fastapi
uvicorn
requests
pydantic